```benchmark                            old ns/op     new ns/op     delta
BenchmarkLoadValues/dbr_10-8         434935        395228        -9.13%
BenchmarkLoadValues/dbr_100-8        720677        733723        +1.81%
BenchmarkLoadValues/dbr_1000-8       3206205       2718942       -15.20%
BenchmarkLoadValues/dbr_10000-8      26411855      18228687      -30.98%
BenchmarkLoadValues/dbr_100000-8     283680737     182544872     -35.65%

benchmark                            old allocs     new allocs     delta
BenchmarkLoadValues/dbr_10-8         239            167            -30.13%
BenchmarkLoadValues/dbr_100-8        2132           1431           -32.88%
BenchmarkLoadValues/dbr_1000-8       21035          14034          -33.28%
BenchmarkLoadValues/dbr_10000-8      210046         140046         -33.33%
BenchmarkLoadValues/dbr_100000-8     2100056        1400055        -33.33%

benchmark                            old bytes     new bytes     delta
BenchmarkLoadValues/dbr_10-8         11662         4645          -60.17%
BenchmarkLoadValues/dbr_100-8        100585        31654         -68.53%
BenchmarkLoadValues/dbr_1000-8       986146        297984        -69.78%
BenchmarkLoadValues/dbr_10000-8      10154700      3274658       -67.75%
BenchmarkLoadValues/dbr_100000-8     101280641     32480560      -67.93%```